pushing everything to github

+++ backend/controllers/todoController.js
@@ -3,7 +3,8 @@ import Todo from '../models/TodoModel.js'
 export const getTodos = async (req, res) => {
     try {
         //get all toods that have an id that matches the autenticatedUser 
-        const todos = await Todo.find({userId:req.params.id})
+    //    console.log(req)
+        const todos = await Todo.find({userId:req.user.uid})
         res.status(200).json(todos)
     } catch(err) {
         console.log(err.message)
@@ -24,6 +25,14 @@ export const createTodo = async (req, res) => {
 
 export const deleteTodo = async (req, res) => {
     try {
+        
+     let result =   await Todo.findById(req.params.id)
+    
+     if(req.user.uid !== result.userId){
+            res.status(401).json({message:"you are not authorized to delete this todo"})
+            return
+     }
+        console.log("🚀 ~ deleteTodo ~ result:", result)
         await Todo.findByIdAndDelete(req.params.id)
         res.status(200).json({ message: 'successfully deleted' })
     } catch (err) {
